reduce document templates to two
