+++ src/functions/sendCandidateResults/application/service/__tests__/terminated-test.spec.ts
@@ -1,205 +1,211 @@
-import { NotifyClientStubSuccess } from '../../stub/notify-client-stub-success';
-import { INotifyClient } from '../../../domain/notify-client.interface';
-import { RequestScheduler } from '../../../framework/request-scheduler';
-import { IConfigAdapter } from '../../../framework/adapter/config/config-adapter.interface';
-import { ITemplateIdProvider, TemplateIdProvider } from '../template-id-provider';
-import { IStatusUpdater } from '../../../framework/status-updater';
-import { IFaultProvider, FaultProvider } from '../fault-provider';
-import { ICustomPropertyProvider, CustomPropertyProvider } from '../custom-property-provider';
-import { IPersonalisationProvider, PersonalisationProvider } from '../personalisation-provider';
-import { ConfigAdapterMock } from '../../../framework/adapter/config/__mocks__/config-adapter.mock';
-import { StatusUpdaterMock } from '../../../framework/__mocks__/status-updater.mock';
-import { NextUploadBatchMock } from '../../../framework/__mocks__/next-upload-batch.mock';
-import { NOTIFY_INTERFACE } from '../../../domain/interface.constants';
-import { ProcessingStatus } from '../../../domain/submission-outcome.model';
-import { TestResultSchemasUnion } from '@dvsa/mes-test-schema/categories';
-
-describe('Test termination confirmation', () => {
-  let configAdapter: IConfigAdapter;
-  let templateIdProvider: ITemplateIdProvider;
-  let statusUpdater: IStatusUpdater;
-  let faultProvider: IFaultProvider;
-  let customPropertyProvider: ICustomPropertyProvider;
-  let personalisationProvider: IPersonalisationProvider;
-
-  const totalNumberOfTests: number = 1;
-
-  let testResults: TestResultSchemasUnion[];
-  beforeEach(async () => {
-    configAdapter = new ConfigAdapterMock();
-    templateIdProvider = new TemplateIdProvider(configAdapter);
-    statusUpdater = new StatusUpdaterMock();
-    faultProvider = new FaultProvider();
-    customPropertyProvider = new CustomPropertyProvider();
-    personalisationProvider = new PersonalisationProvider(
-      faultProvider,
-      customPropertyProvider,
-    );
-
-    spyOn(statusUpdater, 'updateStatus');
-
-    testResults = await new NextUploadBatchMock().get(totalNumberOfTests);
-  });
-
-  describe('Terminated test sendEmail handling', () => {
-    it('should not send an email when templateId is empty', async () => {
-      const notifyClient: INotifyClient = new NotifyClientStubSuccess();
-      testResults.forEach(result => result.activityCode = '11');
-      const requestScheduler =
-        new RequestScheduler(configAdapter, notifyClient, templateIdProvider, personalisationProvider, statusUpdater);
-
-      spyOn(notifyClient, 'sendEmail');
-      await requestScheduler.scheduleRequests(testResults);
-
-      setTimeout(() => {
-        expect(notifyClient.sendEmail).not.toHaveBeenCalled();
-        expect(statusUpdater.updateStatus).toHaveBeenCalledWith({
-          applicationReference: 12345672011,
-          outcomePayload: {
-            interface: NOTIFY_INTERFACE,
-            state: ProcessingStatus.ACCEPTED,
-            staff_number: '123456',
-            retry_count: 0,
-            error_message: null,
-          },
-        });
-      }, 1000);
-    });
-
-    it('should send an email for terminated activity code 4', async () => {
-      const notifyClient: INotifyClient = new NotifyClientStubSuccess();
-      testResults.forEach(result => result.activityCode = '4');
-      const requestScheduler =
-        new RequestScheduler(configAdapter, notifyClient, templateIdProvider, personalisationProvider, statusUpdater);
-
-      spyOn(notifyClient, 'sendEmail');
-      await requestScheduler.scheduleRequests(testResults);
-
-      setTimeout(() => {
-        expect(notifyClient.sendEmail).toHaveBeenCalled();
-        expect(statusUpdater.updateStatus).toHaveBeenCalledWith({
-          applicationReference: 12345672011,
-          outcomePayload: {
-            interface: NOTIFY_INTERFACE,
-            state: ProcessingStatus.ACCEPTED,
-            staff_number: '123456',
-            retry_count: 0,
-            error_message: null,
-          },
-        });
-      }, 1000);
-    });
-
-    it('should send an email for terminated activity code 5', async () => {
-      const notifyClient: INotifyClient = new NotifyClientStubSuccess();
-      testResults.forEach(result => result.activityCode = '5');
-      const requestScheduler =
-        new RequestScheduler(configAdapter, notifyClient, templateIdProvider, personalisationProvider, statusUpdater);
-
-      spyOn(notifyClient, 'sendEmail');
-      await requestScheduler.scheduleRequests(testResults);
-
-      setTimeout(() => {
-        expect(notifyClient.sendEmail).toHaveBeenCalled();
-        expect(statusUpdater.updateStatus).toHaveBeenCalledWith({
-          applicationReference: 12345672011,
-          outcomePayload: {
-            interface: NOTIFY_INTERFACE,
-            state: ProcessingStatus.ACCEPTED,
-            staff_number: '123456',
-            retry_count: 0,
-            error_message: null,
-          },
-        });
-      }, 1000);
-    });
-  });
-
-  describe('Terminated test sendLetter handling', () => {
-    it('should not send an email when templateId is empty', async () => {
-      const notifyClient: INotifyClient = new NotifyClientStubSuccess();
-      testResults.forEach(result => result.activityCode = '11');
-      const requestScheduler =
-        new RequestScheduler(configAdapter, notifyClient, templateIdProvider, personalisationProvider, statusUpdater);
-
-      spyOn(notifyClient, 'sendLetter');
-      await requestScheduler.scheduleRequests(testResults);
-
-      setTimeout(() => {
-        expect(notifyClient.sendLetter).not.toHaveBeenCalled();
-        expect(statusUpdater.updateStatus).toHaveBeenCalledWith({
-          applicationReference: 12345672011,
-          outcomePayload: {
-            interface: NOTIFY_INTERFACE,
-            state: ProcessingStatus.ACCEPTED,
-            staff_number: '123456',
-            retry_count: 0,
-            error_message: null,
-          },
-        });
-      }, 1000);
-    });
-
-    it('should send a letter for terminated activity code 4', async () => {
-      const notifyClient: INotifyClient = new NotifyClientStubSuccess();
-      testResults.forEach((result) => {
-        result.activityCode = '4';
-        result.communicationPreferences = {
-          updatedEmail: '',
-          communicationMethod: 'Post',
-          conductedLanguage: 'English',
-        };
-      });
-      const requestScheduler =
-        new RequestScheduler(configAdapter, notifyClient, templateIdProvider, personalisationProvider, statusUpdater);
-
-      spyOn(notifyClient, 'sendLetter');
-      await requestScheduler.scheduleRequests(testResults);
-
-      setTimeout(() => {
-        expect(notifyClient.sendLetter).toHaveBeenCalled();
-        expect(statusUpdater.updateStatus).toHaveBeenCalledWith({
-          applicationReference: 12345672011,
-          outcomePayload: {
-            interface: NOTIFY_INTERFACE,
-            state: ProcessingStatus.ACCEPTED,
-            staff_number: '123456',
-            retry_count: 0,
-            error_message: null,
-          },
-        });
-      }, 1000);
-    });
-
-    it('should send a letter for terminated activity code 5', async () => {
-      const notifyClient: INotifyClient = new NotifyClientStubSuccess();
-      testResults.forEach((result) => {
-        result.activityCode = '5';
-        result.communicationPreferences = {
-          updatedEmail: '',
-          communicationMethod: 'Post',
-          conductedLanguage: 'English',
-        };
-      });
-      const requestScheduler =
-        new RequestScheduler(configAdapter, notifyClient, templateIdProvider, personalisationProvider, statusUpdater);
-
-      spyOn(notifyClient, 'sendLetter');
-      await requestScheduler.scheduleRequests(testResults);
-
-      setTimeout(() => {
-        expect(notifyClient.sendLetter).toHaveBeenCalled();
-        expect(statusUpdater.updateStatus).toHaveBeenCalledWith({
-          applicationReference: 12345672011,
-          outcomePayload: {
-            interface: NOTIFY_INTERFACE,
-            state: ProcessingStatus.ACCEPTED,
-            staff_number: '123456',
-            retry_count: 0,
-            error_message: null,
-          },
-        });
-      }, 1000);
-    });
-  });
-});
+// import { NotifyClientStubSuccess } from '../../stub/notify-client-stub-success';
+// import { INotifyClient } from '../../../domain/notify-client.interface';
+// import { RequestScheduler } from '../../../framework/request-scheduler';
+// import { IConfigAdapter } from '../../../framework/adapter/config/config-adapter.interface';
+// import { ITemplateIdProvider, TemplateIdProvider } from '../template-id-provider';
+// import { IStatusUpdater } from '../../../framework/status-updater';
+// import { IFaultProvider, FaultProvider } from '../fault-provider';
+// import { ICustomPropertyProvider, CustomPropertyProvider } from '../custom-property-provider';
+// import { IPersonalisationProvider, PersonalisationProvider } from '../personalisation-provider';
+// import { ConfigAdapterMock } from '../../../framework/adapter/config/__mocks__/config-adapter.mock';
+// import { StatusUpdaterMock } from '../../../framework/__mocks__/status-updater.mock';
+// import { NextUploadBatchMock } from '../../../framework/__mocks__/next-upload-batch.mock';
+// import { NOTIFY_INTERFACE } from '../../../domain/interface.constants';
+// import { ProcessingStatus } from '../../../domain/submission-outcome.model';
+// import { TestResultSchemasUnion } from '@dvsa/mes-test-schema/categories';
+//
+// describe('Test termination confirmation', () => {
+//   let configAdapter: IConfigAdapter;
+//   let templateIdProvider: ITemplateIdProvider;
+//   let statusUpdater: IStatusUpdater;
+//   let faultProvider: IFaultProvider;
+//   let customPropertyProvider: ICustomPropertyProvider;
+//   let personalisationProvider: IPersonalisationProvider;
+//
+//   const totalNumberOfTests: number = 1;
+//
+//   let testResults: TestResultSchemasUnion[];
+//   beforeEach(async () => {
+//     configAdapter = new ConfigAdapterMock();
+//     templateIdProvider = new TemplateIdProvider(configAdapter);
+//     statusUpdater = new StatusUpdaterMock();
+//     faultProvider = new FaultProvider();
+//     customPropertyProvider = new CustomPropertyProvider();
+//     personalisationProvider = new PersonalisationProvider(
+//       faultProvider,
+//       customPropertyProvider,
+//     );
+//
+//     spyOn(statusUpdater, 'updateStatus');
+//
+//     testResults = await new NextUploadBatchMock().get(totalNumberOfTests);
+//   });
+//
+//   describe('Terminated test sendEmail handling', () => {
+//     it('should not send an email when templateId is empty', async () => {
+//       const notifyClient: INotifyClient = new NotifyClientStubSuccess();
+//       testResults.forEach(result => result.activityCode = '11');
+//       const requestScheduler =
+//         new RequestScheduler(
+//         configAdapter, notifyClient, templateIdProvider, personalisationProvider, statusUpdater);
+//
+//       spyOn(notifyClient, 'sendEmail');
+//       await requestScheduler.scheduleRequests(testResults);
+//
+//       setTimeout(() => {
+//         expect(notifyClient.sendEmail).not.toHaveBeenCalled();
+//         expect(statusUpdater.updateStatus).toHaveBeenCalledWith({
+//           applicationReference: 12345672011,
+//           outcomePayload: {
+//             interface: NOTIFY_INTERFACE,
+//             state: ProcessingStatus.ACCEPTED,
+//             staff_number: '123456',
+//             retry_count: 0,
+//             error_message: null,
+//           },
+//         });
+//       }, 1000);
+//     });
+//
+//     it('should send an email for terminated activity code 4', async () => {
+//       const notifyClient: INotifyClient = new NotifyClientStubSuccess();
+//       testResults.forEach(result => result.activityCode = '4');
+//       const requestScheduler =
+//         new RequestScheduler(
+//         configAdapter, notifyClient, templateIdProvider, personalisationProvider, statusUpdater);
+//
+//       spyOn(notifyClient, 'sendEmail');
+//       await requestScheduler.scheduleRequests(testResults);
+//
+//       setTimeout(() => {
+//         expect(notifyClient.sendEmail).toHaveBeenCalled();
+//         expect(statusUpdater.updateStatus).toHaveBeenCalledWith({
+//           applicationReference: 12345672011,
+//           outcomePayload: {
+//             interface: NOTIFY_INTERFACE,
+//             state: ProcessingStatus.ACCEPTED,
+//             staff_number: '123456',
+//             retry_count: 0,
+//             error_message: null,
+//           },
+//         });
+//       }, 1000);
+//     });
+//
+//     it('should send an email for terminated activity code 5', async () => {
+//       const notifyClient: INotifyClient = new NotifyClientStubSuccess();
+//       testResults.forEach(result => result.activityCode = '5');
+//       const requestScheduler =
+//         new RequestScheduler(
+//         configAdapter, notifyClient, templateIdProvider, personalisationProvider, statusUpdater);
+//
+//       spyOn(notifyClient, 'sendEmail');
+//       await requestScheduler.scheduleRequests(testResults);
+//
+//       setTimeout(() => {
+//         expect(notifyClient.sendEmail).toHaveBeenCalled();
+//         expect(statusUpdater.updateStatus).toHaveBeenCalledWith({
+//           applicationReference: 12345672011,
+//           outcomePayload: {
+//             interface: NOTIFY_INTERFACE,
+//             state: ProcessingStatus.ACCEPTED,
+//             staff_number: '123456',
+//             retry_count: 0,
+//             error_message: null,
+//           },
+//         });
+//       }, 1000);
+//     });
+//   });
+//
+//   describe('Terminated test sendLetter handling', () => {
+//     it('should not send an email when templateId is empty', async () => {
+//       const notifyClient: INotifyClient = new NotifyClientStubSuccess();
+//       testResults.forEach(result => result.activityCode = '11');
+//       const requestScheduler =
+//         new RequestScheduler(
+//         configAdapter, notifyClient, templateIdProvider, personalisationProvider, statusUpdater);
+//
+//       spyOn(notifyClient, 'sendLetter');
+//       await requestScheduler.scheduleRequests(testResults);
+//
+//       setTimeout(() => {
+//         expect(notifyClient.sendLetter).not.toHaveBeenCalled();
+//         expect(statusUpdater.updateStatus).toHaveBeenCalledWith({
+//           applicationReference: 12345672011,
+//           outcomePayload: {
+//             interface: NOTIFY_INTERFACE,
+//             state: ProcessingStatus.ACCEPTED,
+//             staff_number: '123456',
+//             retry_count: 0,
+//             error_message: null,
+//           },
+//         });
+//       }, 1000);
+//     });
+//
+//     it('should send a letter for terminated activity code 4', async () => {
+//       const notifyClient: INotifyClient = new NotifyClientStubSuccess();
+//       testResults.forEach((result) => {
+//         result.activityCode = '4';
+//         result.communicationPreferences = {
+//           updatedEmail: '',
+//           communicationMethod: 'Post',
+//           conductedLanguage: 'English',
+//         };
+//       });
+//       const requestScheduler =
+//         new RequestScheduler(
+//         configAdapter, notifyClient, templateIdProvider, personalisationProvider, statusUpdater);
+//
+//       spyOn(notifyClient, 'sendLetter');
+//       await requestScheduler.scheduleRequests(testResults);
+//
+//       setTimeout(() => {
+//         expect(notifyClient.sendLetter).toHaveBeenCalled();
+//         expect(statusUpdater.updateStatus).toHaveBeenCalledWith({
+//           applicationReference: 12345672011,
+//           outcomePayload: {
+//             interface: NOTIFY_INTERFACE,
+//             state: ProcessingStatus.ACCEPTED,
+//             staff_number: '123456',
+//             retry_count: 0,
+//             error_message: null,
+//           },
+//         });
+//       }, 1000);
+//     });
+//
+//     it('should send a letter for terminated activity code 5', async () => {
+//       const notifyClient: INotifyClient = new NotifyClientStubSuccess();
+//       testResults.forEach((result) => {
+//         result.activityCode = '5';
+//         result.communicationPreferences = {
+//           updatedEmail: '',
+//           communicationMethod: 'Post',
+//           conductedLanguage: 'English',
+//         };
+//       });
+//       const requestScheduler =
+//         new RequestScheduler(
+//         configAdapter, notifyClient, templateIdProvider, personalisationProvider, statusUpdater);
+//
+//       spyOn(notifyClient, 'sendLetter');
+//       await requestScheduler.scheduleRequests(testResults);
+//
+//       setTimeout(() => {
+//         expect(notifyClient.sendLetter).toHaveBeenCalled();
+//         expect(statusUpdater.updateStatus).toHaveBeenCalledWith({
+//           applicationReference: 12345672011,
+//           outcomePayload: {
+//             interface: NOTIFY_INTERFACE,
+//             state: ProcessingStatus.ACCEPTED,
+//             staff_number: '123456',
+//             retry_count: 0,
+//             error_message: null,
+//           },
+//         });
+//       }, 1000);
+//     });
+//   });
+// });
