+++ src/functions/sendCandidateResults/application/service/test-outcome.ts
@@ -1,0 +1,9 @@
+import { ActivityCode } from '@dvsa/mes-test-schema/categories/common';
+
+export function isPass(activityCode: ActivityCode): boolean {
+  return activityCode === '1';
+}
+
+export function isFail(activityCode: ActivityCode): boolean {
+  return activityCode === '2' || activityCode === '3' || activityCode === '4' || activityCode === '5';
+}
