--- src/functions/sendCandidateResults/domain/communication-type.ts
@@ -1,4 +1,0 @@
-export enum CommunicationType {
-  EMAIL = 'email',
-  LETTER = 'letter',
-}
